Update SeedCount.ijm
Original idea: ungerminated seeds have a certain roundness range, and germinated seeds have lower roundness range, that does not overlap with ungerminated. So, direct analysis in Analyze Particles is prone to false positives/negatives if the seeds aren't "perfect" in the expected roundness range. 
A solution: "Roundness" of each seed is tracked individually, and a drop in roundness is detected and interpreted as germination.